Create Practicals8.py
Practical Title : A program that sorts a list of numbers in
Ascending or Descending order.
Practical Description: The objective of this practical is to
develop a Python program that sorts a list of numbers in Ascending
or Descending order. . This exercise will help students understand
basic programming concepts input/output operations, lists, basic
python methods like a.sort() and a.sort(reverse=True).